LPS-91785 Move table description objects to a more suitable package
