+++ modules/apps/frontend-taglib/frontend-taglib-clay-sample-web/src/main/java/com/liferay/frontend/taglib/clay/sample/web/internal/contributor/SampleTableClayTagContextContributor.java
@@ -15,8 +15,8 @@
 package com.liferay.frontend.taglib.clay.sample.web.internal.contributor;
 
 import com.liferay.frontend.taglib.clay.servlet.taglib.contributor.ClayTagContextContributor;
-import com.liferay.frontend.taglib.clay.servlet.taglib.display.context.table.Field;
-import com.liferay.frontend.taglib.clay.servlet.taglib.display.context.table.Schema;
+import com.liferay.frontend.taglib.clay.servlet.taglib.model.table.Field;
+import com.liferay.frontend.taglib.clay.servlet.taglib.model.table.Schema;
 
 import java.util.Map;
