+++ modules/apps/frontend-taglib/frontend-taglib-clay/bnd.bnd
@@ -6,6 +6,7 @@ Export-Package:\
 	com.liferay.frontend.taglib.clay.servlet.taglib.data,\
 	com.liferay.frontend.taglib.clay.servlet.taglib.display.context,\
 	com.liferay.frontend.taglib.clay.servlet.taglib.display.context.table,\
+	com.liferay.frontend.taglib.clay.servlet.taglib.model.table,\
 	com.liferay.frontend.taglib.clay.servlet.taglib.soy,\
 	com.liferay.frontend.taglib.clay.servlet.taglib.util
 Liferay-JS-Submodules-Bridge:\
